Common command line interface maker
