+++ paltas/Utils/__init__.py
@@ -2,3 +2,4 @@ from . import cosmology_utils
 from . import power_law
 from . import hubble_utils
 from . import lenstronomy_utils
+from . import cli_maker
