+++ paltas/__init__.py
@@ -10,4 +10,4 @@ from . import Sources
 from . import Substructure
 from . import Utils
 from . import MainDeflector
-from . import generate
+from . import generate, robustness_test
